Fixed configs
